Refactor and enhance keyboard and mouse handling
- Introduced modifier key definitions for left and right control, shift, alt, and meta keys.
- Added expanded key handling in the keyboard input processing.
- Implemented mouse cursor movement and button press detection.
- Updated keyboard mapping tables for standard and shifted keys.
- Refactored framebuffer initialization to support ARGB color format.
- Improved pixel manipulation functions for better performance and clarity.
- Adjusted function signatures for pixel operations to use consistent data types.

+++ source/include/ClassiX/framebuf.h
@@ -31,7 +31,7 @@ extern FRAMEBUFFER g_fb;
 
 int init_framebuffer(multiboot_info_t *mbi);
 COLOR get_pixel(uint16_t x, uint16_t y);
-void set_pixel(uint32_t x, uint32_t y, COLOR color);
+void set_pixel(uint16_t x, uint16_t y, COLOR color);
 
 #ifdef __cplusplus
 	}
